refactor(api): try with typescript
